Installed KnpTimeBundle and TwigStringExtra

+++ src/Controller/PinsController.php
@@ -85,10 +85,13 @@ class PinsController extends AbstractController
     /**
      * @Route("/pin/effacer/{id<[0-9]+>}", name="app_pin_delete", methods={"POST"})
      */
-    public function delete(Pin $pin, EntityManagerInterface $em): Response
+    public function delete(Request $request, Pin $pin, EntityManagerInterface $em): Response
     {
-       $em->remove($pin);
-       $em->flush();
-       return $this->redirectToRoute('app_home');
+        //Checks generated CSRF token
+        if($this->isCsrfTokenValid('pin_delete_' . $pin->getId(), $request->request->get('csrf_token'))){
+            $em->remove($pin);
+            $em->flush();
+        }
+        return $this->redirectToRoute('app_home');
     }
 }
